Separate compilation of the note-c lib from Notecard driver
note-c is now included and compiled via newly added Kconfig symbol,
CONFIG_NOTE_C_LIB.

CONFIG_NOTECARD always selects it, as it needs it.

The need for this change is the situation where the developer wants to
test JSON parsing logic in unit test, but doesn't want to include
Notecard driver in the project (as he doesn't need it).