fix: disable i2c logging for both i2c twi and twim
Previously, the code would assert it i2c_nrfx_twi was not being used,
although i2c_nrfx_twim is also a valid option.

We now check for both i2c_nrfx_twi and i2c_nrfx_twim, and disable logging
for the existing option.
If neither is used, we do not disable logging (and do not assert)